refactor: NextAuth 설정 리팩토링 - 콜백/Provider 분리 및 타입 안전성 개선
