+++ apps/web/domain-shared/auth/index.ts
@@ -2,6 +2,12 @@
 export * from "./api/authClient";
 export * from "./api/types";
 
-// Lib
+// Callbacks
+export * from "./lib/callbacks/jwtCallback";
+export * from "./lib/callbacks/signInCallback";
+
+// Utils
 export * from "./lib/appleAuth";
-export * from "./lib/authUtils";
+export * from "./lib/utils";
+export * from "./lib/constants";
+export * from "./lib/providers";
