+++ apps/web/domain-shared/auth/lib/constants.ts
@@ -1,0 +1,8 @@
+export const isProd = process.env.NODE_ENV === "production";
+
+export const secureCookieOptions = {
+  httpOnly: true,
+  sameSite: "lax" as const,
+  path: "/",
+  secure: isProd,
+};
